expand the world
I've come to think that the world is not merely the site of death. It
must also contain beauty--not wishful thinking or idle indulgence but
that reminder of what comes between now and then. Between me and you
lie these words, and perhaps some measure of that quality we call
"humanity." Surely the (human) world needs some of that.