Create caring-for-glaciers_gagné.md
I didn't find this clip online, so I deleted external and source url. It seems this is a clip from a longer podcast found freely here: 

https://poddtoppen.se/podcast/458210899/new-books-in-buddhist-studies/karine-gagne-caring-for-glaciers-land-animals-and-humanity-in-the-himalayas-u-washington-press-2019

Should we post this whole podcast instead?

Also, I think Kashmir and/or Ladakh should be a tag. 

Despite the title, I didn't hear climate change mentioned, though it is a part of the larger story in the author's book, which I briefly mention in the description. Should I put clime-change in the tags?