Create buddhism-trees-and-forests_dhammika.md
This was a toughie. I couldn't find any info or copy online, so I left out the external and source links, just putting in the drive link.

Also, it seems that the title is for a book and the an essay found in that book, so I put this as an essay.

Sorry that I couldn't do more with it. Let me know what I may have missed