Create mulamadhymakakarika_garfield-jay.md
I found this to be an odd one. I understand the 70 verses are selected from a revered monk, but as a document in the library it seems strange. Let's switch this over to Garfield's book which these translations are from?

I couldn't locate a link to this besides the drive. I left external an entry has to have one. Can you point me in the right direction?

Also, is United Kingdom the correct address for the oxford slug?